requirements.txt
google-generativeai>=0.5.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=1.0.0
chromadb>=0.4.0
fastapi>=0.110.0
uvicorn>=0.27.0
streamlit>=1.32.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pypdf>=4.0.0
